+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 
 ### Added
 - Added implementations of `From` to `String` for error enums
+- Added integer categorization methods to `Type` struct